kvserver/rangefeed: separate out buffered registration into its own file
This patch separates out buffered registration into its own file.

Part of: #126560
Release note: none
